gestion_mantenimiento/requirements.txt
Django==4.2.7
gunicorn==21.2.0
dj-database-url==2.1.0
django-crispy-forms==2.1
django-mptt==0.16.0
django-allauth==0.63.3
requests==2.31.0
requests-oauthlib==1.3.1
PyJWT==2.8.0
cryptography==42.0.5
python-dateutil==2.8.2
psycopg2-binary==2.9.9
whitenoise==6.6.0
reportlab==4.0.7
Pillow==10.1.0
# Agrega otras dependencias según tu proyecto
# Instala con: pip install -r requirements.txt